requirements.txt
Django==4.2.4
black==23.7.0
psycopg2-binary==2.9.7
python-decouple==3.0
djangorestframework==3.14.0